pyproject.toml
[project]
name = "gemma3-tool-finetuning"
version = "0.1.0"
description = "使用PEFT微调Gemma3-1b模型以支持工具调用"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "peft>=0.7.0",
    "datasets>=2.14.0",
    "accelerate>=0.20.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "sentencepiece>=0.1.99",
    "gradio>=4.0.0",
    "huggingface_hub>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

# 仅在Linux/Windows上安装bitsandbytes
linux = ["bitsandbytes>=0.41.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
